+++ nextcloud/custom_apps/previewgenerator/lib/SizeHelper.php
@@ -3,32 +3,15 @@
 declare(strict_types=1);
 
 /**
- * @copyright Copyright (c) 2017, Roeland Jago Douma <roeland@famdouma.nl>
- *
- * @author Roeland Jago Douma <roeland@famdouma.nl>
- * @author Richard Steinmetz <richard@steinmetz.cloud>
- *
- * @license AGPL-3.0-or-later
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
+ * SPDX-FileCopyrightText: 2017 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 namespace OCA\PreviewGenerator;
 
 use OCA\PreviewGenerator\AppInfo\Application;
 use OCP\IConfig;
+use OCP\IPreview;
 
 class SizeHelper {
 	public function __construct(
@@ -46,7 +29,8 @@ class SizeHelper {
 
 		$sizes = [
 			'square' => [],
-			'squareUncropped' => [],
+			'fillWidthHeight' => [],
+			'coverWidthHeight' => [],
 			'height' => [],
 			'width' => [],
 		];
@@ -57,7 +41,7 @@ class SizeHelper {
 		$s = 64;
 		while ($s <= $maxW || $s <= $maxH) {
 			$sizes['square'][] = $s;
-			$sizes['squareUncropped'][] = $s;
+			$sizes['fillWidthHeight'][] = $s;
 			$s *= 4;
 		}
 
@@ -78,8 +62,8 @@ class SizeHelper {
 		 * Note that only powers of 4 matter but if users supply different
 		 * stuff it is their own fault and we just ignore it
 		 */
-		$getCustomSizes = function (IConfig $config, $key) {
-			$raw = $config->getAppValue(Application::APP_ID, $key, null);
+		$getCustomSizes = function ($key) {
+			$raw = $this->config->getAppValue(Application::APP_ID, $key, null);
 			if ($raw === null) {
 				return null;
 			}
@@ -99,19 +83,28 @@ class SizeHelper {
 			return $values;
 		};
 
-		$squares = $getCustomSizes($this->config, 'squareSizes');
-		$squaresUncropped = $getCustomSizes($this->config, 'squareUncroppedSizes');
-		$widths = $getCustomSizes($this->config, 'widthSizes');
-		$heights = $getCustomSizes($this->config, 'heightSizes');
+		$squares = $getCustomSizes('squareSizes');
+		$fillWidthHeight = $getCustomSizes('fillWidthHeightSizes')
+			?? $getCustomSizes('squareUncroppedSizes');
+		$coverWidthHeight = $getCustomSizes('coverWidthHeightSizes');
+		$widths = $getCustomSizes('widthSizes');
+		$heights = $getCustomSizes('heightSizes');
 
 		if ($squares !== null) {
 			$sizes['square'] = array_intersect($sizes['square'], $squares);
 		}
 
-		if ($squaresUncropped !== null) {
-			$sizes['squareUncropped'] = array_intersect(
-				$sizes['squareUncropped'],
-				$squaresUncropped,
+		if ($fillWidthHeight !== null) {
+			$sizes['fillWidthHeight'] = array_intersect(
+				$sizes['fillWidthHeight'],
+				$fillWidthHeight,
+			);
+		}
+
+		if ($coverWidthHeight !== null) {
+			$sizes['coverWidthHeight'] = array_filter(
+				$coverWidthHeight,
+				static fn ($size) => $size <= $maxW && $size <= $maxH && self::isPowerOfTwo($size),
 			);
 		}
 
@@ -135,9 +128,22 @@ class SizeHelper {
 			array_map(static function ($squareSize) {
 				return ['width' => $squareSize, 'height' => $squareSize, 'crop' => true];
 			}, $sizes['square']),
-			array_map(static function ($squareSize) {
-				return ['width' => $squareSize, 'height' => $squareSize, 'crop' => false];
-			}, $sizes['squareUncropped']),
+			array_map(static function ($size) {
+				return [
+					'width' => $size,
+					'height' => $size,
+					'crop' => false,
+					'mode' => IPreview::MODE_COVER,
+				];
+			}, $sizes['coverWidthHeight']),
+			array_map(static function ($size) {
+				return [
+					'width' => $size,
+					'height' => $size,
+					'crop' => false,
+					'mode' => IPreview::MODE_FILL,
+				];
+			}, $sizes['fillWidthHeight']),
 			array_map(static function ($heightSize) {
 				return ['width' => -1, 'height' => $heightSize, 'crop' => false];
 			}, $sizes['height']),
@@ -145,5 +151,9 @@ class SizeHelper {
 				return ['width' => $widthSize, 'height' => -1, 'crop' => false];
 			}, $sizes['width'])
 		);
+	}
+
+	private static function isPowerOfTwo(int $n): bool {
+		return ($n & ($n - 1)) === 0;
 	}
 }